separate core and view in users module
